Crate rctl

Resource limits and accounting with RCTL and RACCT

Large parts of this documentation are adapted from the rctl(8) manpage.

Structs

Filter
A filter can match a set of Rules.

Limit
Defines how much of a [Resource] a process can use beofore the defined [Action] triggers.

Rule
A rule represents an [Action] to be taken when a particular [Subject] hits a [Limit] for a [Resource].

RuleParserAdapter
Adapter over objects parseable into a [Rule]

RuleParsingIntoIter
Owning struct implementing IntoIterator, returning a [RuleParserAdapter].

Enums

Action
Represents the action to be taken when a [Subject] offends against a Rule.

Error

ParseError

Resource
An Enum representing a resource type

Signal

State
Enum representing the state of RACCT/RCTL in the kernel.

Subject
A struct representing an RCTL subject.

SubjectType
The type of a [Subject].