Struct rctl::Filter

pub struct Filter { /* fields omitted */ }

A filter can match a set of Rules.

Implementations

impl Filter

pub fn new() -> Filter

Return the filter that matches all rules

Example

let filter = Filter::new();
assert_eq!(filter.to_string(), ":".to_string());

pub fn subject_type(self: Filter, subject_type: &SubjectType) -> Filter

Constrain the filter to a specific [SubjectType]

If the filter is already constrained to a subject, this is a No-Op.

Example

let filter = Filter::new()
    .subject_type(&SubjectType::LoginClass);
assert_eq!(filter.to_string(), "loginclass:".to_string());

pub fn subject(self: Filter, subject: &Subject) -> Filter

Constrain the filter to a specific [Subject]

Resets the subject type.

Example

let filter = Filter::new()
    .subject(&Subject::user_name("nobody").expect("no user 'nobody'"));
assert_eq!(filter.to_string(), "user:nobody".to_string());

pub fn resource(self: Filter, resource: &Resource) -> Filter

Constrain the filter to a specific [Resource]

Example

let filter = Filter::new()
    .resource(&Resource::MemoryLocked);
assert_eq!(filter.to_string(), "::memorylocked".to_string());

pub fn action(self: Filter, action: &Action) -> Filter

Constrain the filter to a specific [Action]

Example

let filter = Filter::new()
    .action(&Action::Deny);
assert_eq!(filter.to_string(), ":::deny".to_string());

pub fn deny(self: Filter) -> Filter

Constrain the filter to the Deny [Action]

Example

let filter = Filter::new()
    .deny();
assert_eq!(filter, Filter::new().action(&Action::Deny));

pub fn log(self: Filter) -> Filter

Constrain the filter to the Log [Action]

Example

let filter = Filter::new()
    .log();
assert_eq!(filter, Filter::new().action(&Action::Log));

pub fn devctl(self: Filter) -> Filter

Constrain the filter to the DevCtl [Action]

Example

let filter = Filter::new()
    .devctl();
assert_eq!(filter, Filter::new().action(&Action::DevCtl));

pub fn signal(self: Filter, signal: Signal) -> Filter

Constrain the filter to the Signal [Action]

Example

let filter = Filter::new()
    .signal(Signal::SIGTERM);
assert_eq!(filter.to_string(), ":::sigterm".to_string());

pub fn limit(self: Filter, limit: &Limit) -> Filter

Constrain the filter to a particular [Limit]

Resets any limit_per, if the given limit has a per set.

Examples

let filter = Filter::new()
    .limit(&Limit::amount(100*1024*1024));
assert_eq!(filter.to_string(), ":::=100m".to_string());

let filter = Filter::new()
    .limit(&Limit::amount_per(100*1024*1024, SubjectType::Process));
assert_eq!(filter.to_string(), ":::=100m/process".to_string());

pub fn to_string(&self) -> String

Return the string representation of the [Filter].

pub fn rules(&self) -> Result<RuleParsingIntoIter<String>, Error>

Return an [IntoIterator] over all Rules matching this [Filter].

Example

List all rules:

use rctl;

let filter = rctl::Filter::new();
for rule in filter.rules() {
    println!("{:?}", rule);
}

pub fn remove_rules(&self) -> Result<(), Error>

Remove all matching Rules from the resource limits database.

Examples

use rctl::{Filter, Subject};
let filter = Filter::new()
    .subject(&Subject::jail_name("testjail_rctl_filter_remove"))
    .remove_rules()
    .expect("Could not remove rules");

Remove all rules, clearing the resource limits database by using the default (':') [Filter]:

use rctl;

rctl::Filter::new().remove_rules().expect("Could not remove all rules");