Struct rctl::Rule

pub struct Rule {
    pub subject: Subject,
    pub resource: Resource,
    pub limit: Limit,
    pub action: Action,
}

A rule represents an [Action] to be taken when a particular [Subject] hits a [Limit] for a [Resource].

Syntax for the string representation of a rule is subject:subject-id:resource:action=amount/per.

Examples

use rctl::{Subject, SubjectType, Resource, Action, Limit, Rule};
let rule = Rule {
    subject: Subject::user_name("nobody").expect("no user 'nobody'"),
    resource: Resource::VMemoryUse,
    action: Action::Deny,
    limit: Limit::amount(1024*1024*1024),
};

assert_eq!(rule.to_string(), "user:nobody:vmemoryuse:deny=1g".to_string());

Fields
subject: Subject resource: Resource limit: Limit action: Action
Methods

impl Rule

pub fn to_string(&self) -> String

pub fn apply(&self) -> Result<(), Error>

Add this rule to the resource limits database.

Example

use rctl::{Rule, Subject, Resource, Action, Limit};

let rule = Rule {
    subject: Subject::jail_name("testjail_rctl_rule_apply_method"),
    resource: Resource::VMemoryUse,
    action: Action::Log,
    limit: Limit::amount(100*1024*1024),
};

rule.apply();

pub fn remove(&self) -> Result<(), Error>

Attempt to remove this rule from the resource limits database.

Example

use rctl::{Rule, Subject, Resource, Action, Limit};

let rule = Rule {
    subject: Subject::jail_name("testjail_rctl_rule_remove_method"),
    resource: Resource::VMemoryUse,
    action: Action::Log,
    limit: Limit::amount(100*1024*1024),
};

rule.remove();