Enum rctl::Action

pub enum Action {
    Deny,
    Log,
    DevCtl,
    Signal(Signal),
    Throttle,
}

Represents the action to be taken when a [Subject] offends against a Rule.

Variants
Deny

Deny the resource allocation

Not supported for the following Resources: CpuTime, Wallclock, ReadBps, WriteBps, ReadIops, WriteIops

Log

Log a warning to the console

DevCtl

Send a notification to devd(8) using system = "RCTL", subsystem = "rule", type = "matched"

Signal(Signal)

Send a signal to the offending process.

Example

use rctl::Signal;
use rctl::Action;

let action = Action::Signal(Signal::SIGTERM);

Throttle

Slow down process execution

Only supported for the following Resources: ReadBps, WriteBps, ReadIops, WriteIops

Methods

impl Action

pub fn as_str(&self) -> &'static str

Return the string representation of the Action according to rctl(8)

Examples

use rctl::Action;
assert_eq!(Action::Deny.as_str(), "deny");

Signals are handled by rctl::Signal:

use rctl::Signal;
assert_eq!(Action::Signal(Signal::SIGKILL).as_str(), "sigkill");