Enum rctl::Subject

pub enum Subject {
    Process(Process),
    Jail(Jail),
    User(User),
    LoginClass(LoginClass),
}

A struct representing an RCTL subject.

From rctl(8):

Subject defines the kind of entity the rule applies to. It can be either process, user, login class, or jail.

Subject ID identifies the subject. It can be user name, numerical user ID login class name, or jail name.

Variants

Process(Process)

Jail(Jail)

User(User)

LoginClass(LoginClass)

Implementations

impl Subject

pub fn process_id(pid: pid_t) -> Self

pub fn user_name(name: &str) -> Result<Self, ParseError>

pub fn user_id(uid: uid_t) -> Self

pub fn login_class<S: Into<String>>(name: S) -> Self

pub fn jail_name<S: Into<String>>(name: S) -> Self

pub fn usage(&self) -> Result<HashMap<Resource, usize>, Error>

Get the resource usage for a specific [Subject].

Example

extern crate libc;

let uid = unsafe { libc::getuid() };
let subject = rctl::Subject::user_id(uid);

let usage = subject.usage()
    .expect("Could not get RCTL usage");

println!("{:#?}", usage);

pub fn limits(&self) -> Result<RuleParsingIntoIter<String>, Error>

Get an IntoIterator over the rules that apply to this subject.